todo/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django==4.2",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
]
